+++ CHANGES.rst
@@ -16,7 +16,7 @@ Changelog
 
 
 **Fixed**
-
+- #991 New client contacts do not have access to their own AR Templates
 - #996 Hide checkbox labels on category expansion
 - #990 Fix client analysisspecs view